finish behavioral sverilog tests
