+++ pymtl3/passes/backends/sverilog/testcases/test_cases.py
@@ -29,34 +29,6 @@ from pymtl3.testcases import add_attributes, \
       CaseInterfaceArrayNonStaticIndexComp, \
       CaseBits32SubCompAttrUpblkComp, CaseBits32ArraySubCompAttrUpblkComp
 
-#-------------------------------------------------------------------------
-# Helper functions that create tv_in and tv_out
-#-------------------------------------------------------------------------
-
-# args: [attr, Bits, idx]
-def _set( *args ):
-  local_dict = {}
-  assert len(args) % 3 == 0
-  _tv_in_str = 'def tv_in( m, tv ):  \n'
-  if len(args) == 0:
-    _tv_in_str += '  pass'
-  for attr, Bits, idx in zip( args[0::3], args[1::3], args[2::3] ):
-    _tv_in_str += f'  m.{attr} = {Bits.__name__}( tv[{idx}] )\n'
-  exec( _tv_in_str, globals(), local_dict )
-  return local_dict['tv_in']
-
-# args: [attr, Bits, idx]
-def _check( *args ):
-  local_dict = {}
-  assert len(args) % 3 == 0
-  _tv_out_str = 'def tv_out( m, tv ):  \n'
-  if len(args) == 0:
-    _tv_out_str += '  pass'
-  for attr, Bits, idx in zip( args[0::3], args[1::3], args[2::3] ):
-    _tv_out_str += f'  assert m.{attr} == {Bits.__name__}( tv[{idx}] )\n'
-  exec( _tv_out_str, globals(), local_dict )
-  return local_dict['tv_out']
-
 CasePassThroughComp = add_attributes( CasePassThroughComp,
     'REF_UPBLK',
     '''\
@@ -64,18 +36,6 @@ CasePassThroughComp = add_attributes( CasePassThroughComp,
           out = in_;
         end
     ''',
-    'TV_IN',
-    _set( 'in_', Bits32, 0 ),
-    'TV_OUT',
-    _check( 'out', Bits32, 1 ),
-    'TEST_VECTOR',
-    [
-        [    0,    0 ],
-        [   42,   42 ],
-        [   24,   24 ],
-        [   -2,   -2 ],
-        [   -1,   -1 ],
-    ]
 )
 
 CaseSequentialPassThroughComp = add_attributes( CaseSequentialPassThroughComp,
@@ -85,18 +45,6 @@ CaseSequentialPassThroughComp = add_attributes( CaseSequentialPassThroughComp,
           out <= in_;
         end
     ''',
-    'TV_IN',
-    _set( 'in_', Bits32, 0 ),
-    'TV_OUT',
-    _check( 'out', Bits32, 1 ),
-    'TEST_VECTOR',
-    [
-        [    0,     0 ],
-        [   42,     0 ],
-        [   24,    42 ],
-        [   -2,    24 ],
-        [   -1,    -2 ],
-    ]
 )
 
 CaseBits32x2ConcatComp = add_attributes( CaseBits32x2ConcatComp,
@@ -106,20 +54,6 @@ CaseBits32x2ConcatComp = add_attributes( CaseBits32x2ConcatComp,
           out = { in_1, in_2 };
         end
     ''',
-    'TV_IN',
-    _set(
-        'in_1', Bits32, 0,
-        'in_2', Bits32, 1,
-    ),
-    'TV_OUT',
-    _check( 'out', Bits64, 2 ),
-    'TEST_VECTOR',
-    [
-        [    0,    0,     concat(    Bits32(0),     Bits32(0) ) ],
-        [   42,    0,     concat(   Bits32(42),     Bits32(0) ) ],
-        [   42,   42,     concat(   Bits32(42),    Bits32(42) ) ],
-        [   -1,   42,     concat(   Bits32(-1),    Bits32(42) ) ],
-    ]
 )
 
 CaseBits32x2ConcatConstComp = add_attributes( CaseBits32x2ConcatConstComp,
@@ -129,14 +63,6 @@ CaseBits32x2ConcatConstComp = add_attributes( CaseBits32x2ConcatConstComp,
           out = { 32'd42, 32'd0 };
         end
     ''',
-    'TV_IN',
-    _set(),
-    'TV_OUT',
-    _check( 'out', Bits64, 0 ),
-    'TEST_VECTOR',
-    [
-        [    concat(    Bits32(42),    Bits32(0) ) ],
-    ]
 )
 
 CaseBits32x2ConcatMixedComp = add_attributes( CaseBits32x2ConcatMixedComp,
@@ -146,17 +72,6 @@ CaseBits32x2ConcatMixedComp = add_attributes( CaseBits32x2ConcatMixedComp,
           out = { in_, 32'd0 };
         end
     ''',
-    'TV_IN',
-    _set( 'in_', Bits32, 0 ),
-    'TV_OUT',
-    _check( 'out', Bits64, 1 ),
-    'TEST_VECTOR',
-    [
-        [  42,  concat(    Bits32(42),    Bits32(0) ) ],
-        [  -1,  concat(    Bits32(-1),    Bits32(0) ) ],
-        [  -2,  concat(    Bits32(-2),    Bits32(0) ) ],
-        [   2,  concat(     Bits32(2),    Bits32(0) ) ],
-    ]
 )
 
 CaseBits64SextInComp = add_attributes( CaseBits64SextInComp,
@@ -166,17 +81,6 @@ CaseBits64SextInComp = add_attributes( CaseBits64SextInComp,
           out = { { 32 { in_[31] } }, in_ };
         end
     ''',
-    'TV_IN',
-    _set( 'in_', Bits32, 0 ),
-    'TV_OUT',
-    _check( 'out', Bits64, 1 ),
-    'TEST_VECTOR',
-    [
-        [  42,   sext(    Bits32(42),    64 ) ],
-        [  -2,   sext(    Bits32(-2),    64 ) ],
-        [  -1,   sext(    Bits32(-1),    64 ) ],
-        [   2,   sext(     Bits32(2),    64 ) ],
-    ]
 )
 
 CaseBits64ZextInComp = add_attributes( CaseBits64ZextInComp,
@@ -186,17 +90,6 @@ CaseBits64ZextInComp = add_attributes( CaseBits64ZextInComp,
           out = { { 32 { 1'b0 } }, in_ };
         end
     ''',
-    'TV_IN',
-    _set( 'in_', Bits32, 0 ),
-    'TV_OUT',
-    _check( 'out', Bits64, 1 ),
-    'TEST_VECTOR',
-    [
-        [  42,   zext(    Bits32(42),    64 ) ],
-        [  -2,   zext(    Bits32(-2),    64 ) ],
-        [  -1,   zext(    Bits32(-1),    64 ) ],
-        [   2,   zext(     Bits32(2),    64 ) ],
-    ]
 )
 
 CaseBits32x2ConcatFreeVarComp = add_attributes( CaseBits32x2ConcatFreeVarComp,
@@ -206,17 +99,6 @@ CaseBits32x2ConcatFreeVarComp = add_attributes( CaseBits32x2ConcatFreeVarComp,
           out = { in_, __const__STATE_IDLE };
         end
     ''',
-    'TV_IN',
-    _set( 'in_', Bits32, 0 ),
-    'TV_OUT',
-    _check( 'out', Bits64, 1 ),
-    'TEST_VECTOR',
-    [
-        [  42,  concat(    Bits32(42),    Bits32(42) ) ],
-        [  -1,  concat(    Bits32(-1),    Bits32(42) ) ],
-        [  -2,  concat(    Bits32(-2),    Bits32(42) ) ],
-        [   2,  concat(     Bits32(2),    Bits32(42) ) ],
-    ]
 )
 
 CaseBits32x2ConcatUnpackedSignalComp = add_attributes( CaseBits32x2ConcatUnpackedSignalComp,
@@ -226,20 +108,6 @@ CaseBits32x2ConcatUnpackedSignalComp = add_attributes( CaseBits32x2ConcatUnpacke
           out = { in_[0], in_[1] };
         end
     ''',
-    'TV_IN',
-    _set(
-        'in_[0]', Bits32, 0,
-        'in_[1]', Bits32, 1,
-    ),
-    'TV_OUT',
-    _check( 'out', Bits64, 2 ),
-    'TEST_VECTOR',
-    [
-        [  42,   2,  concat(    Bits32(42),     Bits32(2) ) ],
-        [  -1,  42,  concat(    Bits32(-1),    Bits32(42) ) ],
-        [  -2,  -1,  concat(    Bits32(-2),    Bits32(-1) ) ],
-        [   2,  -2,  concat(     Bits32(2),    Bits32(-2) ) ],
-    ]
 )
 
 CaseBits32BitSelUpblkComp = add_attributes( CaseBits32BitSelUpblkComp,
@@ -249,17 +117,6 @@ CaseBits32BitSelUpblkComp = add_attributes( CaseBits32BitSelUpblkComp,
           out = in_[1];
         end
     ''',
-    'TV_IN',
-    _set( 'in_', Bits32, 0 ),
-    'TV_OUT',
-    _check( 'out', Bits1, 1 ),
-    'TEST_VECTOR',
-    [
-        [   0,   0 ],
-        [  -1,   1 ],
-        [  -2,   1 ],
-        [   2,   1 ],
-    ]
 )
 
 CaseBits64PartSelUpblkComp = add_attributes( CaseBits64PartSelUpblkComp,
@@ -269,20 +126,6 @@ CaseBits64PartSelUpblkComp = add_attributes( CaseBits64PartSelUpblkComp,
           out = in_[35:4];
         end
     ''',
-    'TV_IN',
-    _set( 'in_', Bits64, 0 ),
-    'TV_OUT',
-    _check( 'out', Bits32, 1 ),
-    'TEST_VECTOR',
-    [
-        [   -1,   -1 ],
-        [   -2,   -1 ],
-        [   -4,   -1 ],
-        [   -8,   -1 ],
-        [  -16,   -1 ],
-        [  -32,   -2 ],
-        [  -64,   -4 ],
-    ]
 )
 
 CaseReducesInx3OutComp = add_attributes( CaseReducesInx3OutComp,
@@ -292,21 +135,6 @@ CaseReducesInx3OutComp = add_attributes( CaseReducesInx3OutComp,
           out = ( ( & in_1 ) & ( | in_2 ) ) | ( ^ in_3 );
         end
     ''',
-    'TV_IN',
-    _set(
-        'in_1', Bits32, 0,
-        'in_2', Bits32, 1,
-        'in_3', Bits32, 2,
-    ),
-    'TV_OUT',
-    _check( 'out', Bits1, 3 ),
-    'TEST_VECTOR',
-    [
-        [  0,   1,    2,  1   ],
-        [ -1,   1,   -1,  1   ],
-        [  9,   8,    7,  1   ],
-        [  9,   8,    0,  0   ],
-    ]
 )
 
 CaseIfBasicComp = add_attributes( CaseIfBasicComp,
@@ -320,19 +148,6 @@ CaseIfBasicComp = add_attributes( CaseIfBasicComp,
             out = 8'd0;
         end
     ''',
-    'TV_IN',
-    _set( 'in_', Bits16, 0 ),
-    'TV_OUT',
-    _check( 'out', Bits8, 1 ),
-    'TEST_VECTOR',
-    [
-        [ 255,   0, ],
-        [  -1, 255, ],
-        [ 511,   1, ],
-        [ 254,   0, ],
-        [  42,   0, ],
-        [   0,   0, ],
-    ]
 )
 
 CaseIfDanglingElseInnerComp = add_attributes( CaseIfDanglingElseInnerComp,
@@ -348,21 +163,6 @@ CaseIfDanglingElseInnerComp = add_attributes( CaseIfDanglingElseInnerComp,
           end
         end
     ''',
-    'TV_IN',
-    _set(
-        'in_1', Bits32, 0,
-        'in_2', Bits32, 1,
-    ),
-    'TV_OUT',
-    _check( 'out', Bits32, 2 ),
-    'TEST_VECTOR',
-    [
-        [    0,    -1,  -1 ],
-        [   42,     0,   0 ],
-        [   24,    42,  42 ],
-        [   -2,    24,  24 ],
-        [   -1,    -2,  -2 ],
-    ]
 )
 
 CaseIfDanglingElseOutterComp = add_attributes( CaseIfDanglingElseOutterComp,
@@ -378,21 +178,6 @@ CaseIfDanglingElseOutterComp = add_attributes( CaseIfDanglingElseOutterComp,
             out = in_2;
         end
     ''',
-    'TV_IN',
-    _set(
-        'in_1', Bits32, 0,
-        'in_2', Bits32, 1,
-    ),
-    'TV_OUT',
-    _check( 'out', Bits32, 2 ),
-    'TEST_VECTOR',
-    [
-        [    0,    -1,   0 ],
-        [   42,     0,   0 ],
-        [   24,    42,   0 ],
-        [   -2,    24,   0 ],
-        [   -1,    -2,   0 ],
-    ]
 )
 
 CaseElifBranchComp = add_attributes( CaseElifBranchComp,
@@ -409,22 +194,6 @@ CaseElifBranchComp = add_attributes( CaseElifBranchComp,
             out = in_3;
         end
     ''',
-    'TV_IN',
-    _set(
-        'in_1', Bits32, 0,
-        'in_2', Bits32, 1,
-        'in_3', Bits32, 2,
-    ),
-    'TV_OUT',
-    _check( 'out', Bits32, 3 ),
-    'TEST_VECTOR',
-    [
-        [    0,    -1,   0,  0 ],
-        [   42,     0,  42, 42 ],
-        [   24,    42,  24, 24 ],
-        [   -2,    24,  -2, -2 ],
-        [   -1,    -2,  -1, -1 ],
-    ]
 )
 
 CaseNestedIfComp = add_attributes( CaseNestedIfComp,
@@ -452,22 +221,6 @@ CaseNestedIfComp = add_attributes( CaseNestedIfComp,
             out = in_1;
         end
     ''',
-    'TV_IN',
-    _set(
-        'in_1', Bits32, 0,
-        'in_2', Bits32, 1,
-        'in_3', Bits32, 2,
-    ),
-    'TV_OUT',
-    _check( 'out', Bits32, 3 ),
-    'TEST_VECTOR',
-    [
-        [    0,    -1,   0,  -1 ],
-        [   42,     0,  42,   0 ],
-        [   24,    42,  24,  42 ],
-        [   -2,    24,  -2,  24 ],
-        [   -1,    -2,  -1,  -2 ],
-    ]
 )
 
 CaseForRangeLowerUpperStepPassThroughComp = add_attributes( CaseForRangeLowerUpperStepPassThroughComp,
@@ -480,30 +233,6 @@ CaseForRangeLowerUpperStepPassThroughComp = add_attributes( CaseForRangeLowerUpp
             out[i] = in_[i];
         end
     ''',
-    'TV_IN',
-    _set(
-        'in_[0]', Bits32, 0,
-        'in_[1]', Bits32, 1,
-        'in_[2]', Bits32, 2,
-        'in_[3]', Bits32, 3,
-        'in_[4]', Bits32, 4,
-    ),
-    'TV_OUT',
-    _check(
-        'out[0]', Bits32, 5,
-        'out[1]', Bits32, 6,
-        'out[2]', Bits32, 7,
-        'out[3]', Bits32, 8,
-        'out[4]', Bits32, 9,
-    ),
-    'TEST_VECTOR',
-    [
-        [    0,    -1,   0,  -1,   0,    0,    -1,   0,  -1,   0, ],
-        [   42,     0,  42,   0,  42,   42,     0,  42,   0,  42, ],
-        [   24,    42,  24,  42,  24,   24,    42,  24,  42,  24, ],
-        [   -2,    24,  -2,  24,  -2,   -2,    24,  -2,  24,  -2, ],
-        [   -1,    -2,  -1,  -2,  -1,   -1,    -2,  -1,  -2,  -1, ],
-    ]
 )
 
 CaseIfExpInForStmtComp = add_attributes( CaseIfExpInForStmtComp,
@@ -514,30 +243,6 @@ CaseIfExpInForStmtComp = add_attributes( CaseIfExpInForStmtComp,
             out[i] = ( i == 1 ) ? in_[i] : in_[0];
         end
     ''',
-    'TV_IN',
-    _set(
-        'in_[0]', Bits32, 0,
-        'in_[1]', Bits32, 1,
-        'in_[2]', Bits32, 2,
-        'in_[3]', Bits32, 3,
-        'in_[4]', Bits32, 4,
-    ),
-    'TV_OUT',
-    _check(
-        'out[0]', Bits32, 5,
-        'out[1]', Bits32, 6,
-        'out[2]', Bits32, 7,
-        'out[3]', Bits32, 8,
-        'out[4]', Bits32, 9,
-    ),
-    'TEST_VECTOR',
-    [
-        [    0,    -1,   0,  -1,   0,    0,    -1,   0,   0,   0, ],
-        [   42,     0,  42,   0,  42,   42,     0,  42,  42,  42, ],
-        [   24,    42,  24,  42,  24,   24,    42,  24,  24,  24, ],
-        [   -2,    24,  -2,  24,  -2,   -2,    24,  -2,  -2,  -2, ],
-        [   -1,    -2,  -1,  -2,  -1,   -1,    -2,  -1,  -1,  -1, ],
-    ]
 )
 
 CaseIfExpUnaryOpInForStmtComp = add_attributes( CaseIfExpUnaryOpInForStmtComp,
@@ -548,30 +253,6 @@ CaseIfExpUnaryOpInForStmtComp = add_attributes( CaseIfExpUnaryOpInForStmtComp,
             out[i] = ( i == 1 ) ? ~in_[i] : in_[0];
         end
     ''',
-    'TV_IN',
-    _set(
-        'in_[0]', Bits32, 0,
-        'in_[1]', Bits32, 1,
-        'in_[2]', Bits32, 2,
-        'in_[3]', Bits32, 3,
-        'in_[4]', Bits32, 4,
-    ),
-    'TV_OUT',
-    _check(
-        'out[0]', Bits32, 5,
-        'out[1]', Bits32, 6,
-        'out[2]', Bits32, 7,
-        'out[3]', Bits32, 8,
-        'out[4]', Bits32, 9,
-    ),
-    'TEST_VECTOR',
-    [
-        [    0,    -1,   0,  -1,   0,    0,    ~-1,   0,   0,   0, ],
-        [   42,     0,  42,   0,  42,   42,     ~0,  42,  42,  42, ],
-        [   24,    42,  24,  42,  24,   24,    ~42,  24,  24,  24, ],
-        [   -2,    24,  -2,  24,  -2,   -2,    ~24,  -2,  -2,  -2, ],
-        [   -1,    -2,  -1,  -2,  -1,   -1,    ~-2,  -1,  -1,  -1, ],
-    ]
 )
 
 CaseIfBoolOpInForStmtComp = add_attributes( CaseIfBoolOpInForStmtComp,
@@ -586,30 +267,6 @@ CaseIfBoolOpInForStmtComp = add_attributes( CaseIfBoolOpInForStmtComp,
               out[i] = 32'd0;
         end
     ''',
-    'TV_IN',
-    _set(
-        'in_[0]', Bits32, 0,
-        'in_[1]', Bits32, 1,
-        'in_[2]', Bits32, 2,
-        'in_[3]', Bits32, 3,
-        'in_[4]', Bits32, 4,
-    ),
-    'TV_OUT',
-    _check(
-        'out[0]', Bits32, 5,
-        'out[1]', Bits32, 6,
-        'out[2]', Bits32, 7,
-        'out[3]', Bits32, 8,
-        'out[4]', Bits32, 9,
-    ),
-    'TEST_VECTOR',
-    [
-        [    0,    -1,   0,  -1,   0,     0,     0,    0,    0,    0, ],
-        [   42,     0,  42,   0,  42,     0,     0,    0,    0,   42, ],
-        [   24,    42,  24,  42,  24,    24,    42,   24,   42,   24, ],
-        [   -2,    24,  -2,  24,  -2,    -2,    24,   -2,   24,   -2, ],
-        [   -1,    -2,  -1,  -2,  -1,    -1,    -2,   -1,   -2,   -1, ],
-    ]
 )
 
 CaseIfTmpVarInForStmtComp = add_attributes( CaseIfTmpVarInForStmtComp,
@@ -626,30 +283,6 @@ CaseIfTmpVarInForStmtComp = add_attributes( CaseIfTmpVarInForStmtComp,
           end
         end
     ''',
-    'TV_IN',
-    _set(
-        'in_[0]', Bits32, 0,
-        'in_[1]', Bits32, 1,
-        'in_[2]', Bits32, 2,
-        'in_[3]', Bits32, 3,
-        'in_[4]', Bits32, 4,
-    ),
-    'TV_OUT',
-    _check(
-        'out[0]', Bits32, 5,
-        'out[1]', Bits32, 6,
-        'out[2]', Bits32, 7,
-        'out[3]', Bits32, 8,
-        'out[4]', Bits32, 9,
-    ),
-    'TEST_VECTOR',
-    [
-        [    0,    -1,   0,  -1,   0,     0,     0,    0,    0,    0, ],
-        [   42,     0,  42,   0,  42,     0,     0,    0,    0,   42, ],
-        [   24,    42,  24,  42,  24,    24,    42,   24,   42,   24, ],
-        [   -2,    24,  -2,  24,  -2,    -2,    24,   -2,   24,   -2, ],
-        [   -1,    -2,  -1,  -2,  -1,    -1,    -2,   -1,   -2,   -1, ],
-    ]
 )
 
 CaseFixedSizeSliceComp = add_attributes( CaseFixedSizeSliceComp,
@@ -660,22 +293,6 @@ CaseFixedSizeSliceComp = add_attributes( CaseFixedSizeSliceComp,
             out[i] = in_[i * 8 +: 8];
         end
     ''',
-    'TV_IN',
-    _set( 'in_', Bits16, 0 ),
-    'TV_OUT',
-    _check(
-        'out[0]', Bits8, 1,
-        'out[1]', Bits8, 2,
-    ),
-    'TEST_VECTOR',
-    [
-        [     -1, 0xff, 0xff ],
-        [      1, 0x01, 0x00 ],
-        [      7, 0x07, 0x00 ],
-        [ 0xff00, 0x00, 0xff ],
-        [ 0x3412, 0x12, 0x34 ],
-        [ 0x9876, 0x76, 0x98 ],
-    ]
 )
 
 CaseBits32FooInBits32OutComp = add_attributes( CaseBits32FooInBits32OutComp,
@@ -685,20 +302,6 @@ CaseBits32FooInBits32OutComp = add_attributes( CaseBits32FooInBits32OutComp,
           out = in_.foo;
         end
     ''',
-    'TV_IN',
-    _set( 'in_', Bits32Foo, 0 ),
-    'TV_OUT',
-    _check( 'out', Bits32, 1 ),
-    'TEST_VECTOR',
-    [
-        [     Bits32Foo(),   0 ],
-        [    Bits32Foo(0),   0 ],
-        [   Bits32Foo(-1),  -1 ],
-        [   Bits32Foo(42),  42 ],
-        [   Bits32Foo(-2),  -2 ],
-        [   Bits32Foo(10),  10 ],
-        [  Bits32Foo(256), 256 ],
-    ]
 )
 
 CaseConstStructInstComp = add_attributes( CaseConstStructInstComp,
@@ -708,14 +311,6 @@ CaseConstStructInstComp = add_attributes( CaseConstStructInstComp,
           out = in_.foo;
         end
     ''',
-    'TV_IN',
-    _set(),
-    'TV_OUT',
-    _check( 'out', Bits32, 0 ),
-    'TEST_VECTOR',
-    [
-        [ 0 ],
-    ]
 )
 
 CaseStructPackedArrayUpblkComp = add_attributes( CaseStructPackedArrayUpblkComp,
@@ -725,18 +320,6 @@ CaseStructPackedArrayUpblkComp = add_attributes( CaseStructPackedArrayUpblkComp,
           out = { in_.foo[0], in_.foo[1], in_.foo[2] };
         end
     ''',
-    'TV_IN',
-    _set( 'in_', Bits32x5Foo, 0 ),
-    'TV_OUT',
-    _check( 'out', Bits96, 1 ),
-    'TEST_VECTOR',
-    [
-        [  Bits32x5Foo([ b32(0), b32(0), b32(0), b32(0),b32(0) ] ), concat( b32(0),   b32(0),   b32(0)  ) ],
-        [  Bits32x5Foo([ b32(-1),b32(-1),b32(-1),b32(0),b32(0) ] ), concat( b32(-1),  b32(-1),  b32(-1) ) ],
-        [  Bits32x5Foo([ b32(42),b32(42),b32(-1),b32(0),b32(0) ] ), concat( b32(42),  b32(42),  b32(-1) ) ],
-        [  Bits32x5Foo([ b32(42),b32(42),b32(42),b32(0),b32(0) ] ), concat( b32(42),  b32(42),  b32(42) ) ],
-        [  Bits32x5Foo([ b32(-1),b32(-1),b32(42),b32(0),b32(0) ] ), concat( b32(-1),  b32(-1),  b32(42) ) ],
-    ]
 )
 
 CaseNestedStructPackedArrayUpblkComp = add_attributes( CaseNestedStructPackedArrayUpblkComp,
@@ -746,18 +329,6 @@ CaseNestedStructPackedArrayUpblkComp = add_attributes( CaseNestedStructPackedArr
           out = { in_.bar[0], in_.woo.foo, in_.foo };
         end
     ''',
-    'TV_IN',
-    _set( 'in_', NestedStructPackedPlusScalar, 0 ),
-    'TV_OUT',
-    _check( 'out', Bits96, 1 ),
-    'TEST_VECTOR',
-    [
-        [   NestedStructPackedPlusScalar(0, [ Bits32(0) , Bits32(0)  ], Bits32Foo(5) ), concat(   Bits32(0), Bits32(5),   Bits32(0) ) ],
-        [  NestedStructPackedPlusScalar(-1, [ Bits32(-1), Bits32(-2) ], Bits32Foo(6) ), concat(  Bits32(-1), Bits32(6),  Bits32(-1) ) ],
-        [  NestedStructPackedPlusScalar(-1, [ Bits32(42), Bits32(43) ], Bits32Foo(7) ), concat(  Bits32(42), Bits32(7),  Bits32(-1) ) ],
-        [  NestedStructPackedPlusScalar(42, [ Bits32(42), Bits32(43) ], Bits32Foo(8) ), concat(  Bits32(42), Bits32(8),  Bits32(42) ) ],
-        [  NestedStructPackedPlusScalar(42, [ Bits32(-1), Bits32(-2) ], Bits32Foo(9) ), concat(  Bits32(-1), Bits32(9),  Bits32(42) ) ],
-    ]
 )
 
 CaseConnectValRdyIfcUpblkComp = add_attributes( CaseConnectValRdyIfcUpblkComp,
@@ -769,25 +340,6 @@ CaseConnectValRdyIfcUpblkComp = add_attributes( CaseConnectValRdyIfcUpblkComp,
           in___rdy = out__rdy;
         end
     ''',
-    'TV_IN',
-    _set(
-        'in_.val', Bits1, 0,
-        'in_.msg', Bits32, 1,
-    ),
-    'TV_OUT',
-    _check(
-        'in_.rdy', Bits1, 0,
-        'out.val', Bits1, 2,
-        'out.msg', Bits32, 3,
-    ),
-    'TEST_VECTOR',
-    [
-        [   1,      0,   1,      0 ],
-        [   0,     42,   0,     42 ],
-        [   1,     42,   1,     42 ],
-        [   1,     -1,   1,     -1 ],
-        [   1,     -2,   1,     -2 ],
-    ]
 )
 
 CaseArrayBits32IfcInUpblkComp = add_attributes( CaseArrayBits32IfcInUpblkComp,
@@ -797,21 +349,6 @@ CaseArrayBits32IfcInUpblkComp = add_attributes( CaseArrayBits32IfcInUpblkComp,
           out = in___1__foo;
         end
     ''',
-    'TV_IN',
-    _set(
-        'in_[0].foo', Bits32, 0,
-        'in_[1].foo', Bits32, 1,
-    ),
-    'TV_OUT',
-    _check( 'out', Bits32, 2 ),
-    'TEST_VECTOR',
-    [
-        [    0,    0,      0 ],
-        [    0,   42,     42 ],
-        [   24,   42,     42 ],
-        [   -2,   -1,     -1 ],
-        [   -1,   -2,     -2 ],
-    ]
 )
 
 CaseBits32SubCompAttrUpblkComp = add_attributes( CaseBits32SubCompAttrUpblkComp,
@@ -821,14 +358,6 @@ CaseBits32SubCompAttrUpblkComp = add_attributes( CaseBits32SubCompAttrUpblkComp,
           out = b__out;
         end
     ''',
-    'TV_IN',
-    _set(),
-    'TV_OUT',
-    _check( 'out', Bits32, 0 ),
-    'TEST_VECTOR',
-    [
-        [ 42 ],
-    ]
 )
 
 CaseBits32ArraySubCompAttrUpblkComp = add_attributes( CaseBits32ArraySubCompAttrUpblkComp,
@@ -838,12 +367,4 @@ CaseBits32ArraySubCompAttrUpblkComp = add_attributes( CaseBits32ArraySubCompAttr
           out = b__1__out;
         end
     ''',
-    'TV_IN',
-    _set(),
-    'TV_OUT',
-    _check( 'out', Bits32, 0 ),
-    'TEST_VECTOR',
-    [
-        [ 42 ],
-    ]
 )
