+++ pymtl3/passes/backends/sverilog/translation/SVTranslator.py
@@ -6,52 +6,10 @@
 """Provide SystemVerilog translator."""
 
 from pymtl3.passes.backends.generic import RTLIRTranslator
+from pymtl3.passes.backends.sverilog.util.utility import sverilog_reserved
 
 from .behavioral import SVBehavioralTranslator as SV_BTranslator
 from .structural import SVStructuralTranslator as SV_STranslator
-
-sverilog_keyword = [
-  # Verilog-1995 reserved keywords
-  "always", "and", "assign", "begin", "buf", "bufif0", "bufif1", "case",
-  "casex", "casez", "cmos", "deassign", "default", "defparam", "disable",
-  "edge", "else", "end", "endcase", "endmodule", "endfunction", "endprimitive",
-  "endspecify", "endtable", "endtask", "event", "for", "force", "forever",
-  "fork", "function", "highz0", "highz1", "if", "ifnone", "initial",
-  "inout", "input", "output", "integer", "join", "large", "macromodule",
-  "medium", "module", "nand", "negedge", "nmos", "nor", "not", "notif0",
-  "notif1", "or", "output", "parameter", "pmos", "posedge", "primitive",
-  "pull0", "pull1", "pullup", "pulldown", "rcmos", "real", "realtime",
-  "reg", "release", "repeat", "rnmos", "rpmos", "rtran", "rtranif0",
-  "rtranif1", "scalared", "small", "specify", "specparam", "strong0",
-  "strong1", "supply0", "supply1", "table", "task", "time", "tran",
-  "tranif0", "tranif1", "tri", "tri0", "tri1", "triand", "trior",
-  "trireg", "vectored", "wait", "wand", "weak0", "weak1", "while",
-  "wire", "wor", "xnor", "xor",
-  # Verilog-2001 reserved keywords
-  "automatic", "cell", "config", "design", "endconfig", "endgenerate",
-  "generate", "genvar", "incdir", "include", "instance", "liblist",
-  "library", "localparam", "noshowcancelled", "pulsestyle_onevent",
-  "pulsestyle_ondetect", "showcancelled", "signed", "unsigned", "use",
-  # Verilog-2005 reserved keywords
-  "uwire",
-  # SystemVerilog-2005 reserved keywords
-  "alias", "always_comb", "always_ff", "always_latch", "assert", "assume",
-  "before", "bind", "bins", "binsof", "bit", "break", "byte", "chandle",
-  "class", "clocking", "const", "constraint", "context", "continue",
-  "cover", "covergroup", "coverpoint", "cross", "dist", "do", "endclass",
-  "endclocking", "endgroup", "endinterface", "endpackage", "endprimitive",
-  "endprogram", "endproperty", "endsequence", "enum", "expect", "export",
-  "extends", "extern", "final", "first_match", "foreach", "forkjoin",
-  "iff", "ignore_bins", "illegal_bins", "import", "inside", "int", "interface",
-  "intersect", "join_any", "join_none", "local", "logic", "longint", "matches",
-  "modport", "new", "null", "package", "packed", "priority", "program",
-  "property", "protected", "pure", "rand", "randc", "randcase", "randsequence",
-  "ref", "return", "sequence", "shortint", "shortreal", "solve", "static",
-  "string", "struct", "super", "tagged", "this", "throughout", "timeprecision",
-  "timeunit", "type", "typedef", "union", "unique", "var", "virtual", "void",
-  "wait_order", "wildcard", "with", "within"
-]
-sverilog_reserved = set( sverilog_keyword )
 
 def mk_SVTranslator( _RTLIRTranslator, _STranslator, _BTranslator ):
