+++ pymtl3/passes/backends/sverilog/translation/structural/test/SVStructuralTranslatorL2_test.py
@@ -5,18 +5,17 @@
 # Date   : May 29, 2019
 """Test the level 2 SystemVerilog structural translator."""
 
-from pymtl3.datatypes import Bits1, Bits32, bitstruct
-from pymtl3.dsl import Component, InPort, OutPort, Wire, connect
-from pymtl3.passes.rtlir import RTLIRDataType as rdt
-from pymtl3.passes.rtlir.util.test_utility import do_test
+import pytest
+
+from pymtl3.passes.backends.sverilog.util.utility import check_eq
+from pymtl3.passes.backends.sverilog.util.test_utility import sverilog_reserved
 
 from ..SVStructuralTranslatorL2 import SVStructuralTranslatorL2
-from .SVStructuralTranslatorL1_test import check_eq, is_sverilog_reserved
 
 
 def local_do_test( m ):
   m.elaborate()
-  SVStructuralTranslatorL2.is_sverilog_reserved = is_sverilog_reserved
+  SVStructuralTranslatorL2.is_sverilog_reserved = lambda x: x in sverilog_reserved
   tr = SVStructuralTranslatorL2( m )
   tr.clear( m )
   tr.translate_structural( m )
