+++ pymtl3/testcases/TestCase.py
@@ -29,3 +29,31 @@ def add_attributes( _cls, *args ):
   for attr, obj in zip( args[::2], args[1::2] ):
     setattr( cls, attr, dedent(obj) if isinstance(obj, str) else obj )
   return cls
+
+#-------------------------------------------------------------------------
+# Helper functions that create tv_in and tv_out
+#-------------------------------------------------------------------------
+
+# args: [attr, Bits, idx]
+def _set( *args ):
+  local_dict = {}
+  assert len(args) % 3 == 0
+  _tv_in_str = 'def tv_in( m, tv ):  \n'
+  if len(args) == 0:
+    _tv_in_str += '  pass'
+  for attr, Bits, idx in zip( args[0::3], args[1::3], args[2::3] ):
+    _tv_in_str += f'  m.{attr} = {Bits.__name__}( tv[{idx}] )\n'
+  exec( _tv_in_str, globals(), local_dict )
+  return local_dict['tv_in']
+
+# args: [attr, Bits, idx]
+def _check( *args ):
+  local_dict = {}
+  assert len(args) % 3 == 0
+  _tv_out_str = 'def tv_out( m, tv ):  \n'
+  if len(args) == 0:
+    _tv_out_str += '  pass'
+  for attr, Bits, idx in zip( args[0::3], args[1::3], args[2::3] ):
+    _tv_out_str += f'  assert m.{attr} == {Bits.__name__}( tv[{idx}] )\n'
+  exec( _tv_out_str, globals(), local_dict )
+  return local_dict['tv_out']
